Topology rendering no longer normalized.
I modified Blob and LPRender so that the various waveforms are no longer normalized since that is what we want when the edge lengths are not the same.  The rendering core still uses some normalization since some of the topology stuff depends on it and I don't want to rewrite everything right now.  I will change it in the future but for now, the blob position is still 0 to 1 but speed, etc is scaled by length.  So when initializing the blob position, you might want to account the length of the edge that you are starting the blob on.