Fix for ShaderToy, Complex vertex shader, Spout Receiver, Kaleidoscopic Fractal
There seems to be some Processing 4 related bug when calling PGraphics.image(srcImg, 0, 0).  It probably has to do with pixel densities but I couldn't get any combination working.  It was also affecting the ShaderToy implementation.  The result was that only the bottom 1/4 of the image would be transferred to the web.  Changing the call to PGraphics.background(srcImg) fixes the issue.

Also includes the nice Complex.vert vertex shader, a Spout receiver for Windows (need to mvn install the jar and copy the dll to natives), and a Kaleidoscopic fractal ShaderToy shader that can use images from the Textures directory).