Azure Kinect + OpenCV blob detection
C++ project and binary for detecting blobs in the depth image via opencv.  Various parameters can be set via the command line or via the sendosc.py script for tuning the sensitive depth and the image margins for blob detection.  Currently the margins are turned up to make testing easier since in WFOV mode it is difficult to find somewhere with enough open space to test properly.  When detected, the blob location will be sent to /spidertrap/blob in LX.